Weapon Switching Groundwork
Made changes to build the groundwork for allowing switching between weapons in inventory and picking weapons up from the ground.

Player now has a script that allows it to control firing rather than having the firing mechanic on each weapon. This should prevent the player from firing every weapon in the scene together.